+++ src/literalizer/__init__.py
@@ -5,10 +5,12 @@ from literalizer._formatters.collection_openers import (
 )
 from literalizer._language import (
     CallStyle,
+    CommandCallStyle,
     CommentConfig,
     DateFormatConfig,
     DatetimeFormatConfig,
     DictFormatConfig,
+    IdentifierCase,
     KeywordCallStyle,
     Language,
     LanguageCls,
@@ -36,11 +38,13 @@ from literalizer._parsing import InputFormat
 __all__ = [
     "BothVariableForms",
     "CallStyle",
+    "CommandCallStyle",
     "CommentConfig",
     "DateFormatConfig",
     "DatetimeFormatConfig",
     "DictFormatConfig",
     "ExistingVariable",
+    "IdentifierCase",
     "InputFormat",
     "KeywordCallStyle",
     "Language",