+++ src/literalizer/languages/common_lisp.py
@@ -40,6 +40,7 @@ from literalizer._language import (
     DictFormatConfig,
     FloatSpecialsMixin,
     HeterogeneousBehavior,
+    IdentifierCase,
     LanguageCls,
     OrderedMapFormatConfig,
     PrefixCallStyle,
@@ -311,6 +312,10 @@ class CommonLisp(metaclass=LanguageCls):
         ERROR = NO_HETEROGENEOUS_BEHAVIOR
 
     heterogeneous_strategies = HeterogeneousStrategies
+
+    identifier_cases: ClassVar[tuple[IdentifierCase, ...]] = (
+        IdentifierCase.KEBAB,
+    )
 
     validate_spec_for_data = no_validate_spec_for_data