+++ src/literalizer/languages/csharp.py
@@ -64,6 +64,7 @@ from literalizer._language import (
     DictFormatConfig,
     FloatSpecialsMixin,
     HeterogeneousBehavior,
+    IdentifierCase,
     KeywordCallStyle,
     LanguageCls,
     ModifierCombination,
@@ -615,6 +616,12 @@ class CSharp(metaclass=LanguageCls):
         ERROR = NO_HETEROGENEOUS_BEHAVIOR
 
     heterogeneous_strategies = HeterogeneousStrategies
+
+    identifier_cases: ClassVar[tuple[IdentifierCase, ...]] = (
+        IdentifierCase.PASCAL,
+        IdentifierCase.CAMEL,
+        IdentifierCase.UPPER_SNAKE,
+    )
 
     modifier_combinations: ClassVar[tuple[ModifierCombination, ...]] = (
         ModifierCombination(