+++ tests/integration/cases/call_ref_args_converted_nonsnake/Lua_call.lua
@@ -1,0 +1,13 @@
+function process(...) end
+local my_var = {
+    1,
+    2,
+    3,
+}
+local my_other = {
+    4,
+    5,
+    6,
+}
+process(my_var, 42)
+process(my_other, 7)