+++ tests/integration/cases/call_ref_args_converted_whole/Cpp_call.cpp
@@ -1,0 +1,12 @@
+#include <initializer_list>
+#include <vector>
+#include <cstddef>
+auto process(auto...) { return 0; }
+void check_() {
+auto my_var = std::vector<int>{
+    1,
+    2,
+    3,
+};
+process(my_var);
+}