+++ tests/integration/cases/call_ref_args_converted_whole/D_call.d
@@ -1,0 +1,10 @@
+import std.json;
+void _check() {
+int process(T...)(T args) { return 0; }
+auto my_var = JSONValue([
+    JSONValue(1),
+    JSONValue(2),
+    JSONValue(3),
+]);
+process(my_var);
+}