+++ tests/integration/cases/call_ref_args_converted_whole/Julia_call.jl
@@ -1,0 +1,7 @@
+process(args...; kwargs...) = nothing
+my_var = [
+    1,
+    2,
+    3,
+]
+process(data=my_var)