+++ tests/integration/cases/call_ref_args_converted_whole/Racket_call.rkt
@@ -1,0 +1,8 @@
+#lang racket
+(define process (make-keyword-procedure (lambda _ (void))))
+(define my-var (list
+    1
+    2
+    3
+))
+(process #:data my-var)