+++ tests/integration/cases/call_wrap_in_file/FSharp_call.fs
@@ -1,0 +1,8 @@
+module Check
+
+let process (_a: obj, _b: obj) : obj = null
+type Val =
+    | FInt of int64
+    | FList of Val list
+process(1, 2)
+process(3, 4)