+++ tests/integration/cases/call_wrap_in_file/R_call.R
@@ -1,0 +1,2 @@
+process(a = 1, b = 2)
+process(a = 3, b = 4)